Replace non-portable hardcoded rpath with meson command
Pylon is typically installed in a non-standard location (/opt/pylon).
This causes the runtime linker to complain about it not finding
the libraries when the plug-in is first loaded. To overcome this one
option is to define the linker's rpath option during build time, which
adds extra paths to the runtime linker search list.

The problem is that the rpath linker option is not portable (will crash
a Window build, for example). Instead of manually hardcoding the rpath
we now rely on meson's install_rpath command which is automatically
ignored when not supported (as with Windows). In these scenarios, however,
the user will require to configure the environment so that the libraries
are found at runtime.