Avoid re-creating pylon gobject when no specific filters are fiven
We cache the device configuration that the user requests in order to check
if later on we need to create a new device, or the user is requesting the same one.
There was an issue in the corner case were no device-index filter was requested (aka -1)
so the logic computed a valid index, and we saved the computed device, instead of the
original requested -1. This caused that every time GStreamer asked for the cam:: child
a new device was created and previous configuration was lost.

Now we save the raw, un-processed values.